initial version of eager loading
this implementation uses the idea from #4 where load the raw data,
lazily apply transformations and then eagerly realize the final
version to reduce the access-time penalty. It might be possible
to refactor this code a bit further to dispatch on the IFD layout
so that I can avoid a lot of these if/else statements and off load
them to compile time.