Don't use `Val` when it can't help
The compiler will specialize for each `N` in `Val{N}`, and so just loading
images with different numbers of slice planes will require additional
compilation. This adds to latency for no benefit, so it's best not to
use `Val` this way.

This throws in a few `@nospecialize` for a couple of other `Val` uses.