Old mmap->lazyio, add new mmap
This PR reworks the handling of "mmapped" images, preserving the
previous implementation and making it accessible with `lazyio=true`.
The main motivation is to take advantage of the OS-provided buffering
of mmapped data, so that elementwise access patterns become much
more efficient, particularly for 3d images. (In some simple demos,
by multiple orders of magnitude.)

I went with a "name swap" because calling the previous implementation
"mmap" is a misnomer: it is a useful form of lazy-loading, but
there was no use of "true" mmapping in the sense of `Mmap.mmap`.

This also introduces the ability to open a mmapped TIFF with `r+`
permissions and thus write (simply by setting values) as well as read.