safe open: call GC before load on Windows (#2)
This attempts to head off a potential bug on Windows: if a file was previously
mmapped in the same session and someone tries to overwrite the file, the presence
of the existing map will trigger a system error. Once can circumvent this if the old map
is garbage-collected. This works only if there are no "live" references to the
memory, but it does not hurt to try.

Perhaps even more importantly, if it fails then an informative warning is displayed.

Co-authored-by: 陈久宁 <chenjiuning@tongyuan.cc>