renamed old mmap to lazyio, implement proper mmapping (#79)
* Old mmap->lazyio, add new mmap

This PR reworks the handling of "mmapped" images, preserving the
previous implementation and making it accessible with `lazyio=true`.
The main motivation is to take advantage of the OS-provided buffering
of mmapped data, so that elementwise access patterns become much
more efficient, particularly for 3d images. (In some simple demos,
by multiple orders of magnitude.)

I went with a "name swap" because calling the previous implementation
"mmap" is a misnomer: it is a useful form of lazy-loading, but
there was no use of "true" mmapping in the sense of `Mmap.mmap`.

This also introduces the ability to open a mmapped TIFF with `r+`
permissions and thus write (simply by setting values) as well as read.

* Add a 3d test

* Fix windows

* Add precompilation

Also fix a bug in sizing of the slice buffer for 1-byte eltypes

* Update docs

* Try again to fix windows

* Try this fix

* Maybe this?

* Another gc

Co-authored-by: Johnny Chen <johnnychen94@hotmail.com>

* Also fix in tests

* add let block (#1)

Co-authored-by: 陈久宁 <chenjiuning@tongyuan.cc>

* fix wording

* Apply suggestions from code review

Co-authored-by: Johnny Chen <johnnychen94@hotmail.com>

* Big doc rewrite

* safe open: call GC before load on Windows (#2)

This attempts to head off a potential bug on Windows: if a file was previously
mmapped in the same session and someone tries to overwrite the file, the presence
of the existing map will trigger a system error. Once can circumvent this if the old map
is garbage-collected. This works only if there are no "live" references to the
memory, but it does not hurt to try.

Perhaps even more importantly, if it fails then an informative warning is displayed.

Co-authored-by: 陈久宁 <chenjiuning@tongyuan.cc>

* DiskTaggedImage -> LazyBufferedTiff

Co-authored-by: Johnny Chen <johnnychen94@hotmail.com>
Co-authored-by: 陈久宁 <chenjiuning@tongyuan.cc>
Co-authored-by: Tamas Nagy <tamas@tamasnagy.com>